+++ src/main/java/com/example/quizmaster/service/QuizService.java
@@ -17,7 +17,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Collections;
@@ -46,7 +45,7 @@ public class QuizService {
                 .build();
 
         quizRepository.save(quiz);
-        return new ApiResponse("Quiz saved successfully", HttpStatus.CREATED);
+        return new ApiResponse("Quiz saved successfully!", HttpStatus.CREATED);
     }
 
     // categoriyalarni olish
@@ -66,13 +65,15 @@ public class QuizService {
                 .data(responseQuizList)
                 .build();
 
-        return new ApiResponse("Quizzes retrieved successfully", HttpStatus.OK, pageable);
+        return new ApiResponse("Quizzes retrieved successfully!", HttpStatus.OK, pageable);
     }
 
     // categoriyani yangilash
     public ApiResponse update(Long id, RequestQuiz requestQuiz) {
-        Quiz quiz = quizRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Quiz not found with id: " + id));
+        Quiz quiz = quizRepository.findById(id).orElse(null);
+        if (quiz == null) {
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
+        }
 
         quiz.setTitle(requestQuiz.getTitle());
         quiz.setDescription(requestQuiz.getDescription());
@@ -80,23 +81,25 @@ public class QuizService {
         quiz.setTimeLimit(requestQuiz.getTimeLimit());
 
         quizRepository.save(quiz);
-        return new ApiResponse("Quiz updated successfully", HttpStatus.OK);
+        return new ApiResponse("Quiz updated successfully!", HttpStatus.OK);
     }
 
     // o'chirish
     public ApiResponse delete(Long id) {
-        Quiz quiz = quizRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Quiz not found with id: " + id));
+        Quiz quiz = quizRepository.findById(id).orElse(null);
+        if (quiz  == null){
+            return new ApiResponse("Quiz not found!",HttpStatus.NOT_FOUND);
+        }
 
         quizRepository.delete(quiz);
-        return new ApiResponse("Quiz deleted successfully", HttpStatus.OK);
+        return new ApiResponse("Quiz deleted successfully!", HttpStatus.OK);
     }
 
     // categoriya boyicha random savollar chiqarish
     public ApiResponse getRandomQuestionsForQuiz(Long quiz) {
         Quiz quiz1 = quizRepository.findById(quiz).orElse(null);
         if (quiz1 == null) {
-            return new ApiResponse("Quiz not found", HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
 
         List<Question> allQuestions = questionRepository.findAllByQuizId(quiz);
@@ -106,17 +109,17 @@ public class QuizService {
         int questionCount = quiz1.getQuestionCount();
 
         if (responseQuestions.size() > questionCount) {
-            return new ApiResponse("Success", HttpStatus.OK, responseQuestions.subList(0, questionCount));
+            return new ApiResponse("Success!", HttpStatus.OK, responseQuestions.subList(0, questionCount));
         }
 
-        return new ApiResponse("Insufficient number of questions available", HttpStatus.CONFLICT,responseQuestions);
+        return new ApiResponse("Insufficient number of questions available!", HttpStatus.CONFLICT,responseQuestions);
     }
 
     // test yechish
-    public ApiResponse passTest(List<ReqPassTest> passTestList, User user, Long quizId) {
+    public ApiResponse passTest(List<ReqPassTest> passTestList, User user, Long quizId,Long timeTaken) {
         Quiz quiz = quizRepository.findById(quizId).orElse(null);
         if (quiz == null) {
-            return new ApiResponse("Quiz not found", HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
 
         LocalDateTime startTime = LocalDateTime.now();
@@ -134,7 +137,6 @@ public class QuizService {
 
         LocalDateTime endTime = LocalDateTime.now();
 
-        long timeTakenInMillis = Duration.between(startTime, endTime).toMillis();
 
         Result result = Result.builder()
                 .correctAnswers((int) correctCountAnswers)
@@ -143,11 +145,11 @@ public class QuizService {
                 .user(user)
                 .startTime(startTime)
                 .endTime(endTime)
-                .timeTaken(timeTakenInMillis)
+                .timeTaken(timeTaken)
                 .build();
 
         resultRepository.save(result);
-        return new ApiResponse("Test passed successfully", HttpStatus.OK);
+        return new ApiResponse("Test passed successfully!", HttpStatus.OK);
     }
 
     // idi boyicha categoruiya qidirish
@@ -155,7 +157,7 @@ public class QuizService {
         Quiz quiz = quizRepository.findById(id).orElse(null);
 
         if (quiz == null) {
-            return new ApiResponse("Quiz not found with this id: " + id, HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
 
         ResponseQuiz responseQuiz = ResponseQuiz.builder()
@@ -167,7 +169,7 @@ public class QuizService {
                 .questionCount(quiz.getQuestionCount())
                 .build();
 
-        return new ApiResponse("Quiz found", HttpStatus.OK, responseQuiz);
+        return new ApiResponse("Quiz found!", HttpStatus.OK, responseQuiz);
 
     }
